requirements.txt
streamlit
google.generativeai
pyttsx3
speech_recognition